# expression: implement vectorized evaluation for `builtinDurationDurationTimeDiffSig`